requirements.txt
Cheetah3 nodig voor html report te genereren

pip install Cheetah3